# internal/gen/http/balance/server/encode_decode.go
// Code generated by goa v3.23.4, DO NOT EDIT.
//
// balance HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/oleshko-g/oggophermart/api/design -o internal/

package server

import (
	"context"
	"errors"
	"net/http"
	"strings"

	balance "github.com/oleshko-g/oggophermart/internal/gen/balance"
	service "github.com/oleshko-g/oggophermart/internal/gen/service"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodePostOrderResponse returns an encoder for responses returned by the
// balance post order endpoint.
func EncodePostOrderResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*balance.PostOrderResult)
		if res.Accepted != nil && *res.Accepted == "yes" {
			w.WriteHeader(http.StatusAccepted)
			return nil
		}
		w.WriteHeader(http.StatusOK)
		return nil
	}
}

// DecodePostOrderRequest returns a decoder for requests sent to the balance
// post order endpoint.
func DecodePostOrderRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*service.JWTToken, error) {
	return func(r *http.Request) (*service.JWTToken, error) {
		var (
			authToken string
			err       error
		)
		authToken = r.Header.Get("Authorization")
		if authToken == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("authToken", "header"))
		}
		if err != nil {
			return nil, err
		}
		payload := NewPostOrderJWTToken(authToken)
		if strings.Contains(payload.AuthToken, " ") {
			// Remove authorization scheme prefix (e.g. "Bearer")
			cred := strings.SplitN(payload.AuthToken, " ", 2)[1]
			payload.AuthToken = cred
		}

		return payload, nil
	}
}

// EncodePostOrderError returns an encoder for errors returned by the post
// order balance endpoint.
func EncodePostOrderError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "Invalid input parameter":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return nil
		case "The order belongs to another user":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusConflict)
			return nil
		case "Invalid order number":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnprocessableEntity)
			return nil
		case "User is not authenticated":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return nil
		case "Internal service error":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return nil
		case "Not implemented":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotImplemented)
			return nil
		default:
			return encodeError(ctx, w, v)
		}
	}
}