# internal/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Goa API
    version: 0.0.1
servers:
    - url: http://localhost:80
      description: Default server for gophermart
paths:
...
    /api/user/orders:
        post:
            tags:
                - balance
            summary: post order balance
            operationId: balance#post order
            responses:
                "200":
                    description: The order has been accepted for processing before.
                "202":
                    description: The order has been accepted for processing.
                "400":
                    description: 'Invalid input parameter: Bad Request response.'
                "401":
                    description: 'User is not authenticated: Unauthorized response.'
                "409":
                    description: 'The order belongs to another user: Conflict response.'
                "422":
                    description: 'Invalid order number: Unprocessable Entity response.'
                "500":
                    description: 'Internal service error: Internal Server Error response.'
                "501":
                    description: 'Not implemented: Not Implemented response.'
            security:
                - jwt_header_Authorization: []
...
components:
...
    securitySchemes:
        jwt_header_Authorization:
            type: http
            description: Secures an endpoint by requiring a valid JWT token.
            scheme: bearer
...